requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
django-multiselectfield==0.1.13
djangorestframework==3.15.2
install==1.3.5
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
